.github/workflows/scripts/Confluence-requirements.txt
atlassian-python-api==4.0.7
pypdf==3.17.4
